+++ src/domain/permissions-description.ts
@@ -1,11 +1,1 @@
-export class PermissionsApp{
-    roles: string[];
-    permissions: string[];
-    
-    constructor(data: any) {
-        this.roles = data.roles;
-        this.permissions = data.permissions;
-    } 
-}
 
-export type PermissionType = "EVENT" | "DIRECTIVE" | "REPORT" | "ADMIN"
