+++ src/views/moduleAdmin/createUss/createUss.tsx
@@ -95,8 +95,6 @@ export const CreateUss = () => {
 
     return (
         <>
-            {/* <Title title={"Crear Usuario"} /> */}
-
         <UserForm
             userForm={user}
             click={isCreate ? createUss : updateUss}
